Refactor: 어드민-클라이언트 파일 분리
- 폴더 구조 수정
