+++ app/domains/streams/admin/schemas.py
